requirements.txt
appdirs==1.4.4
argcomplete==1.10.3
bdfr==2.6.1
beautifulsoup4==4.8.2
boto3==1.26.41
botocore==1.29.41
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
compressed-rtf==1.0.6
cryptography==40.0.1
dict2xml==1.7.2
docx2txt==0.8
ebcdic==1.1.1
elastic-transport==8.4.0
elasticsearch==7.13.0
elasticsearch-dsl==7.4.1
extract-msg==0.28.7
ffmpeg-python==0.2.0
future==0.18.2
google==3.0.0
google-api-core==2.11.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-documentai==2.15.0
google-cloud-vision==3.4.0
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.22.0
idna==3.4
ImageHash==4.3.1
IMAPClient==2.1.0
imgurpython==1.1.7
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.2
mutagen==1.46.0
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
olefile==0.46
packaging==23.0
pdfminer.six==20191110
Pillow==9.3.0
pluginbase==1.0.1
praw==7.7.0
prawcore==2.3.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.16.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scipy==1.9.3
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
textract==1.6.5
TExtractor==0.1.2
toml==0.10.2
tqdm==4.65.0
tzdata==2023.2
tzlocal==4.3
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
websockets==10.4
xlrd==1.2.0
XlsxWriter==3.0.9
yt-dlp==2022.11.11
